Add Tumbler dependency as a private service and update Ristretto
This makes the Flatpak application self-contained from a thumbnail
perspective, as it was already from a configuration perspective. Since
thumbnails are no longer searched in `~/.cache` but in
`$XDG_CACHE_HOME`, and the thumbnailing service name must be
configurable at build time, Ristretto 0.12.2 is needed for this to work
properly.

Changes commited upstream in:
https://gitlab.xfce.org/xfce/tumbler/-/commit/b075248d977a6491a6f0f63f5bc67cdb16c3a0f8
https://gitlab.xfce.org/xfce/tumbler/-/commit/7c7bd60f329e351ebdc02c5268364539d8197d29

A patch has also been added temporarily to Ristretto to overcome a
problem of signal emission from Tumbler. See the message of this patch
and:
https://github.com/flathub/flathub/issues/2770
https://github.com/flathub/org.xfce.ristretto/pull/7